Created Elasticity & Beam Shear Doc/Updated Database
The elasticity & beam shear document has the number from the tables from
the wood handbook required to complete the calculations. I also went
ahead and created all of the calculations in the document. The database
has been updated to reflect the added columns in the database as well as
the database querys.