Changed Code
Updated the database to reflect the new column of beam shear at green
for flat and green. I also updated the properties so that the new
columns would show the affect. I changed the beam deflection code so
that it would be able to know if it was green beam shear or 12 percent.
It appears that the code is working now and we are getting the right
results.